Doc updates for storage (#1570)
