+++ book/src/traits/README.md
@@ -10,31 +10,30 @@ enclaves.
 - [Traits](./traits.md)
 - [External Types](./types.md)
 
-## Using the key store
+## Using storage
 
-The key store is probably one of the most interesting traits because applications
+The store is probably one of the most interesting traits because applications
 that use OpenMLS will interact with it.
-See the [OpenMlsKeyStore trait](./traits.md#openmlskeystore) description for details
-but note that the key used to store, read, and delete values in the key store has
-to be provided as a byte slice.
+See the [StorageProvider trait](./traits.md#storageprovider) description for details.
 
-In the following examples, we have a `ciphersuite` and a `provider` (`OpenMlsCryptoProvider`).
+In the following examples, we have a `ciphersuite` and a `provider` (`OpenMlsProvider`).
 
 ```rust,no_run,noplayground
-{{#include ../../../openmls/tests/key_store.rs:key_store_store}}
+{{#include ../../../openmls/tests/store.rs:store_store}}
+```
+
+Retrieving a value from the store is as simple as calling `read`.
+The retrieved key package bundles the private keys for the init and encryption keys
+as well.
+
+```rust,no_run,noplayground
+{{#include ../../../openmls/tests/store.rs:store_read}}
 ```
 
 The `delete` is called with the identifier to delete a value.
 
 ```rust,no_run,noplayground
-{{#include ../../../openmls/tests/key_store.rs:key_store_delete}}
-```
-
-Retrieving a value from the key store is as simple as calling `read`.
-In this example, we assume we got a `credential` where we want to retrieve the credential bundle, i.e., the private key material.
-
-```rust,no_run,noplayground
-{{#include ../../../openmls/tests/key_store.rs:key_store_read}}
+{{#include ../../../openmls/tests/store.rs:store_delete}}
 ```
 
 [//]: # "links"
